requirements.txt
streamlit

# LangChain Core
langchain==0.3.27
langchain-core==0.3.75


# Google Gemini (PaLM) Integration
langchain-google-genai==2.1.10
google-generativeai

# Hugging Face Integration
langchain-huggingface
transformers
huggingface-hub
sentence-transformers

# Environment Variable Management
python-dotenv

# Machine Learning Utilities
numpy
scikit-learn

youtube-transcript-api
langchain_community
faiss-cpu



